requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
django-crontab==0.7.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
httpie==3.2.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.0.0
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.16.1
PyMySQL==1.1.0
pypng==0.20220715.0
PySocks==1.7.1
python-barcode==0.15.1
python-dateutil==2.8.2
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
